model order and change in cart migration (integer to float)

+++ db/migrate/20221206105855_create_carts.rb
@@ -1,7 +1,7 @@
 class CreateCarts < ActiveRecord::Migration[7.0]
   def change
     create_table :carts do |t|
-      t.integer :total
+      t.float :total
       t.belongs_to :user, index: true
       t.timestamps
     end
